add shared ptr

+++ code/controller/read_write/DNAReader.cpp
@@ -15,7 +15,7 @@ DNAReader::DNAReader(std::string name)
 std::string  DNAReader::read()
 {
     myfile.seekg (0, std::ios::end);
-    int length = myfile.tellg();
+    int length = (int)myfile.tellg();
     myfile.seekg(0);
     char *output = new char[length+1];
     while (!myfile.eof()) {
